requirements.txt
requests==2.19.1
Pillow==6.0.0
matplotlib==3.1.0
numpy==1.16.4
pandas==0.24.2
seaborn==0.9.0
skimage==0.0
scikit_learn==0.21.2
tensorflow==1.14.0
torch==1.1.0
torchvision==0.3.0
